Fix: RouteRepositoryImpl 클래스 getReservationSeats 함수 수정 select에 count() 함수는 long type을 반환하는데 좌석수니까 int로 변경. RouteDetailDto, UserReservedRouteDto도 int로 수정

+++ BE/src/main/java/com/lisade/togeduck/dto/response/UserReservedRouteDto.java
@@ -19,6 +19,6 @@ public class UserReservedRouteDto {
     private String stationName;
     private Integer price;
     private Integer totalSeats;
-    private Long reservedSeats;
+    private Integer reservedSeats;
     private String imagePath;
 }
